+++ packages/example-floors-astro/src/utils/schema.ts
